requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-chroma
langchain-community
langsmith
python-dotenv